refactor(storage): extract _ScanOutputsMixin (insert_opportunity_score + insert_structure_idea)
